Changes for the pull request after review comments
Changes according to review requests
